requirements.txt
requests
selenium
python-dotenv
beautifulsoup4
jellyfish
python-Levenshtein
scikit-learn
Pillow
imagehash
requests